Sun, Mar 5, 2023, 12:12 PM -06:00

+++ runoff/runoff.c
@@ -183,7 +183,7 @@ bool print_winner(void)
     {
         printf("%d\n", candidates[p].votes);
         printf("%d\n", voter_count);
-        float answer = float(candidates[p].votes / voter_count);
+        float answer = double(candidates[p].votes / voter_count);
         if (candidates[p].votes / voter_count > 0.5)
             {
                 printf("%s\n", candidates[p].name);
